Refactor utils into dedicated modules
